3.0.1.4: updated from mtgjson.com to 3.13.1

+++ Mtgdb.Core/StringExtensions.cs
@@ -26,6 +26,12 @@ namespace Mtgdb
 		{
 			var dirInfo = new DirectoryInfo(dir);
 
+			if (!dirInfo.Exists)
+			{
+				dirInfo.Create();
+				return;
+			}
+
 			foreach (var subdirInfo in dirInfo.GetDirectories())
 				subdirInfo.Delete(recursive: true);
